LayoutTests/canvas/philip/tests/2d.line.width.invalid-expected.txt

< [index] >
2d.line.width.invalid
Setting lineWidth to invalid values is ignored
References: 2d.lineWidth.invalid
Actual output:
Passed
